requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
einops==0.8.1
filelock==3.14.0
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jmespath==0.10.0
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.6.2
mmengine==0.10.7
# Editable install with no version control (mmsegmentation==0.29.0)
-e /home/ubuntu/anaconda3/envs/ftan/lib/python3.8/site-packages
model-index==0.1.11
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
prettytable==3.11.0
psutil==7.0.0
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
rich==13.4.2
safetensors==0.5.3
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
thop==0.1.1.post2209072238
timm==1.0.15
tomli==2.2.1
torch==1.11.0+cu115
torchaudio==0.11.0+cu115
torchvision==0.12.0+cu115
tqdm==4.65.2
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
yapf==0.43.0
zipp==3.20.2
